+++ subprojects/common/ddo-core/src/test/java/io/truthencode/ddo/support/AssertionStatus.java
@@ -23,11 +23,11 @@ package io.truthencode.ddo.support;
 public class AssertionStatus {
     /**
      * Convenience method used to determine assertion status.
-     * Used internally to conditionally ignore assertion based unit tests.
+     * Used internally to conditionally ignore assertion-based unit tests.
      *
      * @return true if assertions are enabled.
      */
-    @SuppressWarnings("PMD.DataflowAnomalyAnalysis")
+    @SuppressWarnings({"PMD.DataflowAnomalyAnalysis", "AssignmentExpression", "ReassignedVariable"})
     public static boolean isEnabled() { //    @org.jetbrains.annotations.Contract(pure = true)
         boolean assertOn = false;
         // *assigns* true if assertions are on.