+++ CHANGELOG.md
@@ -9,12 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
--   Deity Feats
--   New Class: Alchemists
+- Deity Feats
+- New Class: Alchemists
 
 ### Changed
 
--   Testing Racial Feats to filter core feats (exclude deity based etc.)
+- Testing Racial Feats to filter core feats (exclude deity based etc.)
 
 ## [0.1.0-alpha] - 2015-12-03